pyproject.toml
[tool.poetry]
name = "lamda-langchain-poetry-docker-demo"
version = "0.1.0"
description = ""
authors = ["Paweł Grabiński <266719@uwr.edu.pl>"]
readme = "README.md"
packages = [{include = "lamda_langchain_poetry_docker_demo"}]

[tool.poetry.dependencies]
python = "^3.8"
aws-cdk-core = "^1.200.0"
aws-cdk-aws-lambda = "^1.200.0"
aws-cdk-aws-apigatewayv2 = "^1.200.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
